ffmpeg: CB1 mpegts seek issue [1/1]
PD#SWPL-67694

Problem:
When seeking in mpegts format, there is no key frame searched. In the
case of very sparse I frames, when seeking to near the end, no I is
really sent to the decoder, causing the picture to be frozen.

Solution:
While seeking to the specified position, it is necessary to seek to the
position where the nearby I frame is located.

Verify:
AndroidR

Signed-off-by: Zhao Yi <zhao.yi@amlogic.com>
Change-Id: I10c6d4dd4061ac37a46677c313f63a7ff5af8f96